+++ projeto2/Apartamento.hpp
@@ -1,7 +1,7 @@
 #ifndef APARTAMENTO_HPP
 #define APARTAMENTO_HPP
 
-#include "imoveis.hpp"
+#include "Imoveis.hpp"
 
 class Apartamento : public Imovel
 {
