--- projeto2/CMakeFiles/Makefile.cmake
@@ -1,46 +1,0 @@
-# CMAKE generated file: DO NOT EDIT!
-# Generated by "Unix Makefiles" Generator, CMake Version 3.10
-
-# The generator used is:
-set(CMAKE_DEPENDS_GENERATOR "Unix Makefiles")
-
-# The top level Makefile was generated from the following files:
-set(CMAKE_MAKEFILE_DEPENDS
-  "CMakeCache.txt"
-  "CMakeFiles/3.10.2/CMakeCCompiler.cmake"
-  "CMakeFiles/3.10.2/CMakeCXXCompiler.cmake"
-  "CMakeFiles/3.10.2/CMakeSystem.cmake"
-  "CMakeLists.txt"
-  "/usr/share/cmake-3.10/Modules/CMakeCInformation.cmake"
-  "/usr/share/cmake-3.10/Modules/CMakeCXXInformation.cmake"
-  "/usr/share/cmake-3.10/Modules/CMakeCommonLanguageInclude.cmake"
-  "/usr/share/cmake-3.10/Modules/CMakeGenericSystem.cmake"
-  "/usr/share/cmake-3.10/Modules/CMakeLanguageInformation.cmake"
-  "/usr/share/cmake-3.10/Modules/CMakeSystemSpecificInformation.cmake"
-  "/usr/share/cmake-3.10/Modules/CMakeSystemSpecificInitialize.cmake"
-  "/usr/share/cmake-3.10/Modules/Compiler/CMakeCommonCompilerMacros.cmake"
-  "/usr/share/cmake-3.10/Modules/Compiler/GNU-C.cmake"
-  "/usr/share/cmake-3.10/Modules/Compiler/GNU-CXX.cmake"
-  "/usr/share/cmake-3.10/Modules/Compiler/GNU.cmake"
-  "/usr/share/cmake-3.10/Modules/Platform/Linux-GNU-C.cmake"
-  "/usr/share/cmake-3.10/Modules/Platform/Linux-GNU-CXX.cmake"
-  "/usr/share/cmake-3.10/Modules/Platform/Linux-GNU.cmake"
-  "/usr/share/cmake-3.10/Modules/Platform/Linux.cmake"
-  "/usr/share/cmake-3.10/Modules/Platform/UnixPaths.cmake"
-  )
-
-# The corresponding makefile is:
-set(CMAKE_MAKEFILE_OUTPUTS
-  "Makefile"
-  "CMakeFiles/cmake.check_cache"
-  )
-
-# Byproducts of CMake generate step:
-set(CMAKE_MAKEFILE_PRODUCTS
-  "CMakeFiles/CMakeDirectoryInformation.cmake"
-  )
-
-# Dependency information for all targets:
-set(CMAKE_DEPEND_INFO_FILES
-  "CMakeFiles/pj2.dir/DependInfo.cmake"
-  )
