--- projeto2/CMakeFiles/Progress/count.txt
@@ -1,1 +1,0 @@
-4
