--- projeto2/CMakeFiles/pj2.dir/CXX.includecache
@@ -1,46 +1,0 @@
-#IncludeRegexLine: ^[ 	]*[#%][ 	]*(include|import)[ 	]*[<"]([^">]+)([">])
-
-#IncludeRegexScan: ^.*$
-
-#IncludeRegexComplain: ^$
-
-#IncludeRegexTransform: 
-
-/home/carlos/Documentos/Coding/cppTests/lp1/lab-p1/projeto2/SistemaImobiliaria.cpp
-SistemaImobiliaria.hpp
-/home/carlos/Documentos/Coding/cppTests/lp1/lab-p1/projeto2/SistemaImobiliaria.hpp
-
-/home/carlos/Documentos/Coding/cppTests/lp1/lab-p1/projeto2/SistemaImobiliaria.hpp
-list
--
-vector
--
-imoveis.hpp
-/home/carlos/Documentos/Coding/cppTests/lp1/lab-p1/projeto2/imoveis.hpp
-
-/home/carlos/Documentos/Coding/cppTests/lp1/lab-p1/projeto2/imoveis.cpp
-imoveis.hpp
-/home/carlos/Documentos/Coding/cppTests/lp1/lab-p1/projeto2/imoveis.hpp
-string
--
-iostream
--
-
-/home/carlos/Documentos/Coding/cppTests/lp1/lab-p1/projeto2/imoveis.hpp
-string
--
-
-/home/carlos/Documentos/Coding/cppTests/lp1/lab-p1/projeto2/main.cpp
-iostream
--
-string
--
-list
--
-vector
--
-imoveis.hpp
-/home/carlos/Documentos/Coding/cppTests/lp1/lab-p1/projeto2/imoveis.hpp
-SistemaImobiliaria.hpp
-/home/carlos/Documentos/Coding/cppTests/lp1/lab-p1/projeto2/SistemaImobiliaria.hpp
-
