--- projeto2/CMakeFiles/pj2.dir/DependInfo.cmake
@@ -1,22 +1,0 @@
-# The set of languages for which implicit dependencies are needed:
-set(CMAKE_DEPENDS_LANGUAGES
-  "CXX"
-  )
-# The set of files for implicit dependencies of each language:
-set(CMAKE_DEPENDS_CHECK_CXX
-  "/home/carlos/Documentos/Coding/cppTests/lp1/lab-p1/projeto2/SistemaImobiliaria.cpp" "/home/carlos/Documentos/Coding/cppTests/lp1/lab-p1/projeto2/CMakeFiles/pj2.dir/SistemaImobiliaria.cpp.o"
-  "/home/carlos/Documentos/Coding/cppTests/lp1/lab-p1/projeto2/imoveis.cpp" "/home/carlos/Documentos/Coding/cppTests/lp1/lab-p1/projeto2/CMakeFiles/pj2.dir/imoveis.cpp.o"
-  "/home/carlos/Documentos/Coding/cppTests/lp1/lab-p1/projeto2/main.cpp" "/home/carlos/Documentos/Coding/cppTests/lp1/lab-p1/projeto2/CMakeFiles/pj2.dir/main.cpp.o"
-  )
-set(CMAKE_CXX_COMPILER_ID "GNU")
-
-# The include file search paths:
-set(CMAKE_CXX_TARGET_INCLUDE_PATH
-  )
-
-# Targets to which this target links.
-set(CMAKE_TARGET_LINKED_INFO_FILES
-  )
-
-# Fortran module output directory.
-set(CMAKE_Fortran_TARGET_MODULE_DIR "")
