--- projeto2/CMakeFiles/pj2.dir/cmake_clean.cmake
@@ -1,12 +1,0 @@
-file(REMOVE_RECURSE
-  "CMakeFiles/pj2.dir/main.cpp.o"
-  "CMakeFiles/pj2.dir/imoveis.cpp.o"
-  "CMakeFiles/pj2.dir/SistemaImobiliaria.cpp.o"
-  "pj2.pdb"
-  "pj2"
-)
-
-# Per-language clean rules from dependency scanning.
-foreach(lang CXX)
-  include(CMakeFiles/pj2.dir/cmake_clean_${lang}.cmake OPTIONAL)
-endforeach()
