--- projeto2/CMakeFiles/pj2.dir/depend.internal
@@ -1,14 +1,0 @@
-# CMAKE generated file: DO NOT EDIT!
-# Generated by "Unix Makefiles" Generator, CMake Version 3.10
-
-CMakeFiles/pj2.dir/SistemaImobiliaria.cpp.o
- /home/carlos/Documentos/Coding/cppTests/lp1/lab-p1/projeto2/SistemaImobiliaria.cpp
- /home/carlos/Documentos/Coding/cppTests/lp1/lab-p1/projeto2/SistemaImobiliaria.hpp
- /home/carlos/Documentos/Coding/cppTests/lp1/lab-p1/projeto2/imoveis.hpp
-CMakeFiles/pj2.dir/imoveis.cpp.o
- /home/carlos/Documentos/Coding/cppTests/lp1/lab-p1/projeto2/imoveis.cpp
- /home/carlos/Documentos/Coding/cppTests/lp1/lab-p1/projeto2/imoveis.hpp
-CMakeFiles/pj2.dir/main.cpp.o
- /home/carlos/Documentos/Coding/cppTests/lp1/lab-p1/projeto2/SistemaImobiliaria.hpp
- /home/carlos/Documentos/Coding/cppTests/lp1/lab-p1/projeto2/imoveis.hpp
- /home/carlos/Documentos/Coding/cppTests/lp1/lab-p1/projeto2/main.cpp
