--- projeto2/CMakeFiles/pj2.dir/depend.make
@@ -1,14 +1,0 @@
-# CMAKE generated file: DO NOT EDIT!
-# Generated by "Unix Makefiles" Generator, CMake Version 3.10
-
-CMakeFiles/pj2.dir/SistemaImobiliaria.cpp.o: SistemaImobiliaria.cpp
-CMakeFiles/pj2.dir/SistemaImobiliaria.cpp.o: SistemaImobiliaria.hpp
-CMakeFiles/pj2.dir/SistemaImobiliaria.cpp.o: imoveis.hpp
-
-CMakeFiles/pj2.dir/imoveis.cpp.o: imoveis.cpp
-CMakeFiles/pj2.dir/imoveis.cpp.o: imoveis.hpp
-
-CMakeFiles/pj2.dir/main.cpp.o: SistemaImobiliaria.hpp
-CMakeFiles/pj2.dir/main.cpp.o: imoveis.hpp
-CMakeFiles/pj2.dir/main.cpp.o: main.cpp
-
