--- projeto2/CMakeFiles/pj2.dir/flags.make
@@ -1,10 +1,0 @@
-# CMAKE generated file: DO NOT EDIT!
-# Generated by "Unix Makefiles" Generator, CMake Version 3.10
-
-# compile CXX with /usr/bin/c++
-CXX_FLAGS = -std=c++14 -g  
-
-CXX_DEFINES = 
-
-CXX_INCLUDES = 
-
