--- projeto2/CMakeFiles/pj2.dir/link.txt
@@ -1,1 +1,0 @@
-/usr/bin/c++  -std=c++14 -g  -rdynamic CMakeFiles/pj2.dir/main.cpp.o CMakeFiles/pj2.dir/imoveis.cpp.o CMakeFiles/pj2.dir/SistemaImobiliaria.cpp.o  -o pj2
