--- projeto2/CMakeFiles/pj2.dir/progress.make
@@ -1,5 +1,0 @@
-CMAKE_PROGRESS_1 = 1
-CMAKE_PROGRESS_2 = 2
-CMAKE_PROGRESS_3 = 3
-CMAKE_PROGRESS_4 = 4
-
