--- projeto2/CMakeFiles/progress.marks
@@ -1,1 +1,0 @@
-4
