+++ projeto2/Casa.cpp
@@ -1,23 +1,21 @@
-#ifndef APARTAMENTO_CPP
-#define APARTAMENTO_CPP
+#ifndef CASA_CPP
+#define CASA_CPP
 
-#include "apartamento.hpp"
+#include "Casa.hpp"
 #include <string>
 #include <iostream>
 
-int Apartamento::getAndar() { return andar; }
-int Apartamento::getNQuartos() { return nQuartos; }
-int Apartamento::getNVagas() { return nVagas; }
-double Apartamento::getArea() { return area; }
-char Apartamento::getPosicao() { return posicao; }
+int Casa::getNPavimentos() { return nPavimentos; }
+int Casa::getNQuartos() { return nQuartos; }
+double Casa::getAreaTerreno() { return areaTerreno; }
+double Casa::getAreaConstruida() { return areaConstruida; }
 
-void Apartamento::setAndar(int a) { andar = a; }
-void Apartamento::setNQuartos(int nQ) { nQuartos = nQ; }
-void Apartamento::setNVagas(int nV) { nVagas = nV; }
-void Apartamento::setArea(double a) { area = a; }
-void Apartamento::setPosicao(char p) { posicao = p; }
+void Casa::setNPavimentos(int nP) { nPavimentos = nP; }
+void Casa::setNQuartos(int nQ) { nQuartos = nQ; }
+void Casa::setAreaTerreno(double aT) { areaTerreno = aT; }
+void Casa::setAreaConstruida(double aC) { areaConstruida = aC; }
 
-void Apartamento::cadastrar()
+void Casa::cadastrar()
 {
     std::cout << "Informe o Titulo: ";
     std::cin >> tituloAnuncio;
@@ -43,23 +41,20 @@ void Apartamento::cadastrar()
     std::cout << "Informe a disponibilidade: ";
     std::cin >> disponibilidade;
 
-    std::cout << "Informe o andar: ";
-    std::cin >> andar;
+    std::cout << "Informe o numero de pavimentos: ";
+    std::cin >> nPavimentos;
 
     std::cout << "Informe o numero de quartos: ";
     std::cin >> nQuartos;
 
-    std::cout << "Informe o numero de vagas de garagem: ";
-    std::cin >> nVagas;
+    std::cout << "Informe a area do terreno: ";
+    std::cin >> areaTerreno;
 
-    std::cout << "Informe a area: ";
-    std::cin >> area;
-
-    std::cout << "Informe a posicao: ";
-    std::cin >> posicao;
+    std::cout << "Informe a area construida: ";
+    std::cin >> areaConstruida;
 }
 
-void Apartamento::exibir()
+void Casa::exibir()
 {
     std::cout << "Titulo: " << tituloAnuncio << std::endl;
     std::cout << "Rua: " << rua << std::endl;
@@ -70,10 +65,10 @@ void Apartamento::exibir()
     std::cout << "Valor: " << valor << std::endl;
     std::cout << "Disponibilidade: " << disponibilidade << std::endl;
 
-    std::cout << "Andar: " << andar << std::endl;
+    std::cout << "Numero de pavimentos: " << nPavimentos << std::endl;
     std::cout << "Numero de quartos: " << nQuartos << std::endl;
-    std::cout << "Numero de vagas de garagem: " << nVagas << std::endl;
-    std::cout << "Area: " << area << std::endl;
-    std::cout << "Posicao: " << posicao << std::endl;
+    std::cout << "Area do terreno: " << areaTerreno << std::endl;
+    std::cout << "Area construida: " << areaConstruida << std::endl;
 }
+
 #endif
