+++ projeto2/Casa.hpp
@@ -1,7 +1,7 @@
 #ifndef CASA_HPP
 #define CASA_HPP
 
-#include "imoveis.hpp"
+#include "Imoveis.hpp"
