+++ projeto2/GerenteDePersistencia.cpp
@@ -1,0 +1,7 @@
+#ifndef GERENTEDEPERSISTENCIA_CPP
+#define GERENTEDEPERSISTENCIA_CPP
+
+#include "GerenteDePersistencia.hpp"
+
+
+#endif
