+++ projeto2/GerenteDePersistencia.hpp
@@ -1,0 +1,16 @@
+#ifndef GERENTEDEPERSISTENCIA_HPP
+#define GERENTEDEPERSISTENCIA_HPP
+
+#include <list>
+#include <vector>
+#include "Imoveis.hpp"
+
+class GerenteDePersistencia
+{
+    public:
+        std::vector<Imovel*> recuperaListaImoveis();
+        void salvaListaImoveis();
+};
+
+
+#endif
