--- projeto2/Makefile
@@ -1,238 +1,0 @@
-# CMAKE generated file: DO NOT EDIT!
-# Generated by "Unix Makefiles" Generator, CMake Version 3.10
-
-# Default target executed when no arguments are given to make.
-default_target: all
-
-.PHONY : default_target
-
-# Allow only one "make -f Makefile2" at a time, but pass parallelism.
-.NOTPARALLEL:
-
-
-#=============================================================================
-# Special targets provided by cmake.
-
-# Disable implicit rules so canonical targets will work.
-.SUFFIXES:
-
-
-# Remove some rules from gmake that .SUFFIXES does not remove.
-SUFFIXES =
-
-.SUFFIXES: .hpux_make_needs_suffix_list
-
-
-# Suppress display of executed commands.
-$(VERBOSE).SILENT:
-
-
-# A target that is always out of date.
-cmake_force:
-
-.PHONY : cmake_force
-
-#=============================================================================
-# Set environment variables for the build.
-
-# The shell in which to execute make rules.
-SHELL = /bin/sh
-
-# The CMake executable.
-CMAKE_COMMAND = /usr/bin/cmake
-
-# The command to remove a file.
-RM = /usr/bin/cmake -E remove -f
-
-# Escaping for special characters.
-EQUALS = =
-
-# The top-level source directory on which CMake was run.
-CMAKE_SOURCE_DIR = /home/carlos/Documentos/Coding/cppTests/lp1/lab-p1/projeto2
-
-# The top-level build directory on which CMake was run.
-CMAKE_BINARY_DIR = /home/carlos/Documentos/Coding/cppTests/lp1/lab-p1/projeto2
-
-#=============================================================================
-# Targets provided globally by CMake.
-
-# Special rule for the target rebuild_cache
-rebuild_cache:
-	@$(CMAKE_COMMAND) -E cmake_echo_color --switch=$(COLOR) --cyan "Running CMake to regenerate build system..."
-	/usr/bin/cmake -H$(CMAKE_SOURCE_DIR) -B$(CMAKE_BINARY_DIR)
-.PHONY : rebuild_cache
-
-# Special rule for the target rebuild_cache
-rebuild_cache/fast: rebuild_cache
-
-.PHONY : rebuild_cache/fast
-
-# Special rule for the target edit_cache
-edit_cache:
-	@$(CMAKE_COMMAND) -E cmake_echo_color --switch=$(COLOR) --cyan "No interactive CMake dialog available..."
-	/usr/bin/cmake -E echo No\ interactive\ CMake\ dialog\ available.
-.PHONY : edit_cache
-
-# Special rule for the target edit_cache
-edit_cache/fast: edit_cache
-
-.PHONY : edit_cache/fast
-
-# The main all target
-all: cmake_check_build_system
-	$(CMAKE_COMMAND) -E cmake_progress_start /home/carlos/Documentos/Coding/cppTests/lp1/lab-p1/projeto2/CMakeFiles /home/carlos/Documentos/Coding/cppTests/lp1/lab-p1/projeto2/CMakeFiles/progress.marks
-	$(MAKE) -f CMakeFiles/Makefile2 all
-	$(CMAKE_COMMAND) -E cmake_progress_start /home/carlos/Documentos/Coding/cppTests/lp1/lab-p1/projeto2/CMakeFiles 0
-.PHONY : all
-
-# The main clean target
-clean:
-	$(MAKE) -f CMakeFiles/Makefile2 clean
-.PHONY : clean
-
-# The main clean target
-clean/fast: clean
-
-.PHONY : clean/fast
-
-# Prepare targets for installation.
-preinstall: all
-	$(MAKE) -f CMakeFiles/Makefile2 preinstall
-.PHONY : preinstall
-
-# Prepare targets for installation.
-preinstall/fast:
-	$(MAKE) -f CMakeFiles/Makefile2 preinstall
-.PHONY : preinstall/fast
-
-# clear depends
-depend:
-	$(CMAKE_COMMAND) -H$(CMAKE_SOURCE_DIR) -B$(CMAKE_BINARY_DIR) --check-build-system CMakeFiles/Makefile.cmake 1
-.PHONY : depend
-
-#=============================================================================
-# Target rules for targets named pj2
-
-# Build rule for target.
-pj2: cmake_check_build_system
-	$(MAKE) -f CMakeFiles/Makefile2 pj2
-.PHONY : pj2
-
-# fast build rule for target.
-pj2/fast:
-	$(MAKE) -f CMakeFiles/pj2.dir/build.make CMakeFiles/pj2.dir/build
-.PHONY : pj2/fast
-
-SistemaImobiliaria.o: SistemaImobiliaria.cpp.o
-
-.PHONY : SistemaImobiliaria.o
-
-# target to build an object file
-SistemaImobiliaria.cpp.o:
-	$(MAKE) -f CMakeFiles/pj2.dir/build.make CMakeFiles/pj2.dir/SistemaImobiliaria.cpp.o
-.PHONY : SistemaImobiliaria.cpp.o
-
-SistemaImobiliaria.i: SistemaImobiliaria.cpp.i
-
-.PHONY : SistemaImobiliaria.i
-
-# target to preprocess a source file
-SistemaImobiliaria.cpp.i:
-	$(MAKE) -f CMakeFiles/pj2.dir/build.make CMakeFiles/pj2.dir/SistemaImobiliaria.cpp.i
-.PHONY : SistemaImobiliaria.cpp.i
-
-SistemaImobiliaria.s: SistemaImobiliaria.cpp.s
-
-.PHONY : SistemaImobiliaria.s
-
-# target to generate assembly for a file
-SistemaImobiliaria.cpp.s:
-	$(MAKE) -f CMakeFiles/pj2.dir/build.make CMakeFiles/pj2.dir/SistemaImobiliaria.cpp.s
-.PHONY : SistemaImobiliaria.cpp.s
-
-imoveis.o: imoveis.cpp.o
-
-.PHONY : imoveis.o
-
-# target to build an object file
-imoveis.cpp.o:
-	$(MAKE) -f CMakeFiles/pj2.dir/build.make CMakeFiles/pj2.dir/imoveis.cpp.o
-.PHONY : imoveis.cpp.o
-
-imoveis.i: imoveis.cpp.i
-
-.PHONY : imoveis.i
-
-# target to preprocess a source file
-imoveis.cpp.i:
-	$(MAKE) -f CMakeFiles/pj2.dir/build.make CMakeFiles/pj2.dir/imoveis.cpp.i
-.PHONY : imoveis.cpp.i
-
-imoveis.s: imoveis.cpp.s
-
-.PHONY : imoveis.s
-
-# target to generate assembly for a file
-imoveis.cpp.s:
-	$(MAKE) -f CMakeFiles/pj2.dir/build.make CMakeFiles/pj2.dir/imoveis.cpp.s
-.PHONY : imoveis.cpp.s
-
-main.o: main.cpp.o
-
-.PHONY : main.o
-
-# target to build an object file
-main.cpp.o:
-	$(MAKE) -f CMakeFiles/pj2.dir/build.make CMakeFiles/pj2.dir/main.cpp.o
-.PHONY : main.cpp.o
-
-main.i: main.cpp.i
-
-.PHONY : main.i
-
-# target to preprocess a source file
-main.cpp.i:
-	$(MAKE) -f CMakeFiles/pj2.dir/build.make CMakeFiles/pj2.dir/main.cpp.i
-.PHONY : main.cpp.i
-
-main.s: main.cpp.s
-
-.PHONY : main.s
-
-# target to generate assembly for a file
-main.cpp.s:
-	$(MAKE) -f CMakeFiles/pj2.dir/build.make CMakeFiles/pj2.dir/main.cpp.s
-.PHONY : main.cpp.s
-
-# Help Target
-help:
-	@echo "The following are some of the valid targets for this Makefile:"
-	@echo "... all (the default if no target is provided)"
-	@echo "... clean"
-	@echo "... depend"
-	@echo "... rebuild_cache"
-	@echo "... pj2"
-	@echo "... edit_cache"
-	@echo "... SistemaImobiliaria.o"
-	@echo "... SistemaImobiliaria.i"
-	@echo "... SistemaImobiliaria.s"
-	@echo "... imoveis.o"
-	@echo "... imoveis.i"
-	@echo "... imoveis.s"
-	@echo "... main.o"
-	@echo "... main.i"
-	@echo "... main.s"
-.PHONY : help
-
-
-
-#=============================================================================
-# Special targets to cleanup operation of make.
-
-# Special rule to run CMake to check the build system integrity.
-# No rule that depends on this can have commands that come from listfiles
-# because they might be regenerated.
-cmake_check_build_system:
-	$(CMAKE_COMMAND) -H$(CMAKE_SOURCE_DIR) -B$(CMAKE_BINARY_DIR) --check-build-system CMakeFiles/Makefile.cmake 0
-.PHONY : cmake_check_build_system
-
